+++ crypto/math/src/fft/mod.rs
@@ -2,12 +2,12 @@ pub mod bit_reversing;
 #[cfg(feature = "alloc")]
 pub mod bowers_fft;
 pub mod errors;
-pub mod fft;
 #[cfg(feature = "alloc")]
 pub mod roots_of_unity;
 
-#[cfg(all(test, feature = "alloc"))]
-mod bowers_fft_tests;
+/// Reference radix-2 FFT, used only to cross-check the production Bowers FFT.
+#[cfg(test)]
+pub(crate) mod reference_fft;
 
 #[cfg(all(test, feature = "alloc"))]
 pub(crate) mod test_helpers;
