+++ crypto/math/src/tests/bowers_fft_tests.rs
@@ -2,9 +2,9 @@
 //!
 //! Separated from the main implementation to keep the library code concise.
 
-use super::bowers_fft::*;
 use crate::fft::bit_reversing::in_place_bit_reverse_permute;
-use crate::fft::fft::in_place_nr_2radix_fft;
+use crate::fft::bowers_fft::*;
+use crate::fft::reference_fft::in_place_nr_2radix_fft;
 use crate::fft::roots_of_unity::get_powers_of_primitive_root;
 use crate::field::element::FieldElement;
 use crate::field::goldilocks::GoldilocksField;
