+++ crypto/math/src/tests/mod.rs
@@ -1,4 +1,5 @@
 pub mod barycentric_tests;
+pub mod bowers_fft_tests;
 pub mod fft_friendly_extensions_goldilocks_tests;
 pub mod fft_friendly_u64_goldilocks_tests;
 pub mod fft_tests;
